+++ ModelSwapperSkins/BoneMapping/InitializerRules/BoneInitializerRules_Lemurian.cs
@@ -25,16 +25,25 @@ namespace ModelSwapperSkins.BoneMapping.InitializerRules
                 {
                     case BoneType.Stomach:
                     case BoneType.Chest:
-                    case BoneType.ArmLowerL:
-                    case BoneType.ArmLowerR:
                         bone.RotationOffset *= Quaternion.Euler(0f, 180f, 0f);
                         break;
                     case BoneType.Head:
-                        bone.RotationOffset *= Quaternion.Euler(-45f, 0f, 0f);
+                        bone.RotationOffset *= Quaternion.Euler(270f, 0f, 0f);
+                        break;
+                    case BoneType.Jaw:
+                        bone.RotationOffset *= Quaternion.Euler(330f, 0f, 0f);
                         break;
                     case BoneType.ArmUpperL:
+                        bone.PositionOffset += new Vector3(0f, -0.15f, 0f);
+                        bone.RotationOffset *= Quaternion.Euler(0f, 90f, 0f);
+                        break;
                     case BoneType.ArmUpperR:
                         bone.PositionOffset += new Vector3(0f, -0.15f, 0f);
+                        bone.RotationOffset *= Quaternion.Euler(0f, 270f, 0f);
+                        break;
+                    case BoneType.Toe1L:
+                    case BoneType.Toe1R:
+                        bone.RotationOffset *= Quaternion.Euler(0f, 180f, 0f);
                         break;
                 }