+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   * Imp Overlord model matching
   * Jellyfish model matching
   * Larva model matching
+  * Lemurian body part rotation
   * Model cut-offs persisting when model is used as a skin
 
 * Removed R2API.Language dependency